# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Register a shop + owner and auto-login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Log in with login + password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Rotate refresh token for a new token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Current user + shop",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password (revokes all sessions)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/accessories": {
      "get": {
        "operationId": "AccessoriesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Match on name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "description": "true → only quantity > 0, false → only quantity = 0",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List accessories",
        "tags": [
          "accessories"
        ]
      },
      "post": {
        "operationId": "AccessoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccessoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an accessory + first stock entry",
        "tags": [
          "accessories"
        ]
      }
    },
    "/accessories/{id}": {
      "get": {
        "operationId": "AccessoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one accessory",
        "tags": [
          "accessories"
        ]
      },
      "patch": {
        "operationId": "AccessoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccessoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an accessory (not quantity)",
        "tags": [
          "accessories"
        ]
      },
      "delete": {
        "operationId": "AccessoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete an accessory",
        "tags": [
          "accessories"
        ]
      }
    },
    "/accessories/{id}/stock": {
      "post": {
        "operationId": "AccessoriesController_addStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add stock (new intake at possibly new cost)",
        "tags": [
          "accessories"
        ]
      },
      "get": {
        "operationId": "AccessoriesController_listStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockEntryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List stock intake history",
        "tags": [
          "accessories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Phone & Accessory Shop API",
    "description": "Backend for a small retail shop: phones, accessories, sales, debts, expenses and statistics. Multi-tenant — all data scoped by shop.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "shopName": {
            "type": "string",
            "example": "Aziz Mobile",
            "description": "Do'kon nomi"
          },
          "fullName": {
            "type": "string",
            "example": "Abdulaziz Karimov"
          },
          "login": {
            "type": "string",
            "example": "aziz.mobile",
            "description": "3–32 chars, letters/digits/._- only, globally unique"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "supersecret"
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8,
            "example": "supersecret"
          },
          "inviteCode": {
            "type": "string",
            "description": "Required only when the server sets INVITE_CODE"
          }
        },
        "required": [
          "shopName",
          "fullName",
          "login",
          "password",
          "confirmPassword"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token TTL in seconds"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "example": "aziz.mobile"
          },
          "password": {
            "type": "string",
            "example": "supersecret"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The raw refresh token issued at login/refresh"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token to revoke"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "fullName",
          "login",
          "role",
          "isActive"
        ]
      },
      "MeShopDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "object",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "labelFooter": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "phone",
          "labelFooter"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/MeUserDto"
          },
          "shop": {
            "$ref": "#/components/schemas/MeShopDto"
          }
        },
        "required": [
          "user",
          "shop"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword",
          "confirmPassword"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAccessoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "USB-C kabel 1m"
          },
          "purchasePrice": {
            "type": "number",
            "example": 15000
          },
          "quantity": {
            "type": "number",
            "example": 10,
            "description": "Initial stock quantity"
          },
          "salePrice": {
            "type": "number",
            "example": 25000
          },
          "imageUrl": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "purchasePrice",
          "quantity"
        ]
      },
      "AccessoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "purchasePrice": {
            "type": "number"
          },
          "salePrice": {
            "type": "object",
            "nullable": true
          },
          "quantity": {
            "type": "number"
          },
          "imageUrl": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "purchasePrice",
          "quantity",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAccessoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "USB-C kabel 1m"
          },
          "purchasePrice": {
            "type": "number",
            "example": 15000
          },
          "salePrice": {
            "type": "number",
            "example": 25000
          },
          "imageUrl": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "AddStockDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 20,
            "description": "Quantity to add"
          },
          "purchasePrice": {
            "type": "number",
            "example": 16000,
            "description": "Cost of this intake"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "quantity",
          "purchasePrice"
        ]
      },
      "StockEntryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "purchasePrice": {
            "type": "number"
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "quantity",
          "purchasePrice",
          "createdAt"
        ]
      },
...
    }
  }
}